pyproject.toml
[project]
name = "agent-platform"
version = "0.1.0"
description = "Minimal agent platform with provider routing, registry, telemetry and load testing."
readme = "README.md"
requires-python = ">=3.14,<3.15"
dependencies = [
    "fastapi>=0.135.3",
    "uvicorn[standard]>=0.35.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.43",
    "aiosqlite>=0.21.0",
    "prometheus-client>=0.22.0",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-sdk>=1.36.0",
    "opentelemetry-exporter-otlp>=1.36.0",
    "mlflow>=3.11.1",
    "tiktoken>=0.12.0",
    "litellm>=1.76.0",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.15.10",
    "ty>=0.0.1a17",
    "pre-commit>=4.5.1",
    "locust>=2.43.4",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py314"
extend-exclude = ["opposite-platform", "agentbeats-tau2"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
minversion = "9.0"
addopts = "-q --strict-markers --strict-config"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ty.src]
include = ["shared", "services", "tests"]
